refactor(code, mcp-common): drop the Ifc suffix and a vestigial Spring property
Two independent cleanups from the review in #642.

Name the two interfaces in code.format plainly, the way every other
interface in the repository is named (CommandRunner, McpTool,
ColumnarDataSource, ViolationSink): FormatterIfc becomes Formatter and
ImportRemoverIfc becomes ImportRemover. Formatter was taken by the
implementation, which is now EclipseFormatter after the Eclipse JDT
CodeFormatter it wraps -- descriptive on the implementation side, the way
UnusedImportsRemover already was. The Ifc suffix that protogen emits into
its *generated* builder interfaces is a separate, deliberate convention
(Utils.IFC_SUFFIX) and is left alone, so no ArchUnit rule bans the suffix
outright.

Remove spring.main.allow-bean-definition-overriding=true from the data and
context servers. Nothing in AbstractMcpConfiguration needs it: the two
McpSyncServer @Bean methods carry different bean names, and their
conditions are mutually exclusive anyway, because
McpStreamableServerTransportProvider extends McpServerTransportProviderBase
rather than McpServerTransportProvider, so @ConditionalOnBean never matches
in streamable-http mode. adopt has always run without the property. The
transport ITs pass over all three transports with it gone. Enabling
override globally suppressed a real class of wiring error for every bean in
the context, not just the one it was added for.

A new unit test pins the SDK hierarchy the removal rests on, so an SDK
release that widened it fails here rather than quietly defining two sync
servers for streamable-http to choose between.
