+++ code/context/src/main/resources/application.properties
@@ -4,7 +4,6 @@ server.port=8082
 # endpoint has no authentication, so it must not be reachable from the network.
 # Override deliberately (and add authentication) before exposing it more widely.
 server.address=127.0.0.1
-spring.main.allow-bean-definition-overriding=true
 
 # Confine the project scanners to these directories (File.pathSeparator-separated
 # list of absolute paths). Any requested path is resolved to its real location
