--- code/protogen-maven-plugin/src/main/java/io/github/adamw7/tools/code/format/FormatterIfc.java
@@ -1,6 +1,0 @@
-package io.github.adamw7.tools.code.format;
-
-public interface FormatterIfc {
-
-	String format(String code);
-}
