+++ data/src/main/resources/application.properties
@@ -4,7 +4,6 @@ server.port=8081
 # endpoint has no authentication, so it must not be reachable from the network.
 # Override deliberately (and add authentication) before exposing it more widely.
 server.address=127.0.0.1
-spring.main.allow-bean-definition-overriding=true
 
 logging.level.root=INFO
 logging.level.org.springframework.boot.autoconfigure=ERROR
